refactor(container): organize calengrade feature
